Refactor admin settings management page
- Reorganized settings management into a more modular structure with dedicated components for better maintainability.
- Introduced `CategorySection` and `SettingCard` components to enhance the UI and improve the separation of concerns.
- Utilized a custom hook `useAdminSettings` for managing settings state and fetching logic.
- Updated settings configuration to be imported from a constants file for easier management.
- Enhanced loading and saving states for individual settings to improve user experience.
